Add GitHub Actions workflow to build Unifont BMP TTF
The Unifont project has decided to stop distributing precompiled
TrueType fonts for the Unicode Basic Multilingual Plane (BMP) starting
with version 15.1.01. The recommended migration path is to use OpenType
fonts instead, but modifying software to accommodate this new format can
be costly. So if we want to continue using TTF fonts for the foreseeable
future, we need to build them ourselves from source, following the
official instructions.

To make this process easy, reproducible, and convenient, this PR
introduces a GitHub Actions workflow that can be manually triggered for
this effect, requesting the Unifont version to be built. When the
workflow completes, an artifact containing the built TTF file is
uploaded.

The described workflow automatically verifies that the GPG signature of
the Unifont source tarball checks up with the contents of
`unifoundry-keys.gpg`, which can be downloaded from
https://unifoundry.com/1A09227B1F435A33_public.asc, according to the
official signature verification instructions.